Reorganize classes inside modules.
